Move getProducts to hocs utils
