+++ assets/js/blocks/products/utils.js
@@ -4,7 +4,6 @@
 import { __ } from '@wordpress/i18n';
 import { Button, Placeholder } from '@wordpress/components';
 import classNames from 'classnames';
-import apiFetch from '@wordpress/api-fetch';
 
 /**
  * Internal dependencies
@@ -67,27 +66,3 @@ export const renderHiddenContentPlaceholder = ( blockTitle, blockIcon ) => (
 		) }
 	</Placeholder>
 );
-
-export const getProducts = ( queryArgs ) => {
-	const args = {
-		catalog_visibility: 'visible',
-		status: 'publish',
-		...queryArgs,
-	};
-	return apiFetch( {
-		path:
-			'/wc/blocks/products?' +
-			Object.entries( args )
-				.map( ( arg ) => arg.join( '=' ) )
-				.join( '&' ),
-		parse: false,
-	} ).then( ( response ) => {
-		return response.json().then( ( products ) => {
-			const totalProducts = parseInt(
-				response.headers.get( 'x-wp-total' ),
-				10
-			);
-			return { products, totalProducts };
-		} );
-	} );
-};
